pyproject.toml
[project]
name = "langchain-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "arxiv>=2.4.1",
    "ddgs>=9.11.4",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.115.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn>=0.34.0",
    "wikipedia>=1.4.0",
]
